Art
"Art is a diverse range of human activities in creating visual, auditory or performing artifacts (artworks), expressing the author's imaginative, conceptual idea, or technical skill, intended to be appreciated for their beauty or emotional power.[1][2] In their most general form these activities include the production of works of art, the criticism of art, the study of the history of art, and the aesthetic dissemination of art.

Music, theatre, film, dance, and other performing arts, as well as literature and other media such as interactive media, are included in a broader definition of art or the arts.[1][3] Until the 17th century, art referred to any skill or mastery and was not differentiated from crafts or sciences. In modern usage after the 17th century, where aesthetic considerations are paramount, the fine arts are separated and distinguished from acquired skills in general, such as the decorative or applied arts.

Though the definition of what constitutes art is disputed[4][5][6] and has changed over time, general descriptions mention an idea of imaginative or technical skill stemming from human agency[7] and creation.[8] The nature of art and related concepts, such as creativity and interpretation, are explored in a branch of philosophy known as aesthetics.[9]"
Source:  https://en.wikipedia.org/wiki/Art