A kind of working network structure (#1)
This project is now able to create a network and train it based on learning data! It still has problems learning "the binding of isaac: afterbirth+" but it is easily able to learn a XOR gate and calculate the right outcome.

* I tried to add neurons

This is a sketch of a possible implementation of neurons into a neural network. I'm not sure that this will work at the end.

* Renamed the api namespace to os!

The api namespace will be specific for GPU processing APIs. The os namespace includes all classes that are OS specific

* Added PaCo: everything about Parallel Computing

I like PaCo as an abbreviation.

* New PaCo classes and finally a workeing network.... well kind of

It all seams to work... and is slow AF.

* Improved image scaling performance

* Created a network for BOI: Afterbirth+

The network is very slow at the moment, the next tasks should be to rework the network.
The biggest time problem seems to be the bitmap processing, I'm also looking into ways to improve this.

* Changed the way of defining paco activation functions

* Removed the NeuronGroup classes

All neurons are now normal and inside the NeuronLayer class.

And 'O' saves the last screenshot.

* fixed includes

* Multithreaded screen capturing

I moved the screen capturing into a different thread this improved performance... and destroyed my will to live. The crashes I had to deal with... OMG. The A.I. is now up to 24 - 29 UPS.

Some stats:
~10 Total PC crashes
~30 Runntime errors
~3 Drinks
~8 Hours for something so simple
~1 will to live was lost
~0 People will ever read this
~1 Github was taken over by Microsoft
~14 Hours playing BOI

* The ScreenCapturer can now easily run by a different thread.

This makes the code way cleaner and easier

* Fixed the capture area offset.

* [WIP]Implemented / reworked the bitmap struct

I'm not sure if I'm going to keep it as int bitmaps or if I switch back do floats, but the basic organisation will stay

More stats:
+1  Glass of Wine
+3 unnecessary statistics
+1  Stayed up to late for this

* Changed the new implemented bitmap to a float bitmap.

Not to much more just some big plans and bug fixes.

* Reworked the float bitmap struct

Cuda is no longer required to run this network. The bitmap utils are now moved into the BitmapUtil file in the paco directory.

* Updated some formatting :D

* Changed some includes

* Something, well I tried to implement Backpropagation... and failed

* Art

"Art is a diverse range of human activities in creating visual, auditory or performing artifacts (artworks), expressing the author's imaginative, conceptual idea, or technical skill, intended to be appreciated for their beauty or emotional power.[1][2] In their most general form these activities include the production of works of art, the criticism of art, the study of the history of art, and the aesthetic dissemination of art.

Music, theatre, film, dance, and other performing arts, as well as literature and other media such as interactive media, are included in a broader definition of art or the arts.[1][3] Until the 17th century, art referred to any skill or mastery and was not differentiated from crafts or sciences. In modern usage after the 17th century, where aesthetic considerations are paramount, the fine arts are separated and distinguished from acquired skills in general, such as the decorative or applied arts.

Though the definition of what constitutes art is disputed[4][5][6] and has changed over time, general descriptions mention an idea of imaginative or technical skill stemming from human agency[7] and creation.[8] The nature of art and related concepts, such as creativity and interpretation, are explored in a branch of philosophy known as aesthetics.[9]"
Source:  https://en.wikipedia.org/wiki/Art

* [More than broken] aka "Art 2"

Let's rewrite the code for the fourth time!

* New sum functions for the matrix

* [Broken] Working test network

* [BROken] copied another network and the test is working now

* [Still broken] Addes backpropagation for the bias :D

At least I think so, it works at least!

* [Broken] I did some refactoring

I converted the test network to use matrices instead of neuron classes

* Added a byte buffer struct and refactored a bit

https://i.imgur.com/JOlBUHx.jpg

* Reworked the matrix save and load functions

* Reworked the matrix save as text functions

Nothing more, it's not like someone will ever read this. Maybe I should post some interesting information her so someone one dark day will find this when they search for the information on google. Sounds like a Plan... Now I just need to find something interesting... DUX, next commit!!

* Finally a build that is not completely broken

It does something. I don't know what but it seems to actually process the frames and give me a valid output without crashing

* Started to implement back propagation

* Some Testing

And this definitely needs some work :/